intent/src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
#
anyconfig==0.9.11         # via kedro
appdirs==1.4.4            # via black
appnope==0.1.0            # via ipykernel, ipython
arrow==0.15.7             # via jinja2-time
attrs==19.3.0             # via black, jsonschema, pytest
backcall==0.2.0           # via ipython
binaryornot==0.4.4        # via cookiecutter
black==v19.10b0           # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
bleach==3.1.5             # via nbconvert
cachetools==4.1.1         # via kedro
certifi==2020.6.20        # via requests
chardet==3.0.4            # via binaryornot, requests
click==7.1.2              # via black, cookiecutter, kedro, pip-tools
cookiecutter==1.6.0       # via kedro
coverage==5.1             # via pytest-cov
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
flake8==3.8.3             # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
fsspec==0.6.3             # via kedro
future==0.18.2            # via cookiecutter
idna==2.10                # via requests
importlib-metadata==1.7.0  # via flake8, jsonschema, pluggy, pytest
ipykernel==5.3.0          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via jupyterlab, nbformat, notebook, qtconsole, traitlets
ipython==7.16.1           # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in, ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
isort==4.3.21             # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
jedi==0.17.1              # via ipython
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.11.2            # via cookiecutter, jinja2-time, nbconvert, notebook
jmespath==0.10.0          # via kedro
jsonschema==3.2.0         # via jupyterlab-launcher, nbformat
jupyter-client==5.3.5     # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in, ipykernel, jupyter-console, kedro, notebook, qtconsole
jupyter-console==6.1.0    # via jupyter
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
jupyterlab-launcher==0.10.5  # via jupyterlab
jupyterlab==0.31.1        # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
kedro==0.16.2             # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mistune==0.8.4            # via nbconvert
more-itertools==8.4.0     # via pytest
nbconvert==5.6.1          # via jupyter, notebook
nbformat==5.0.7           # via ipywidgets, nbconvert, nbstripout, notebook
nbstripout==0.3.3         # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
notebook==6.0.3           # via jupyter, jupyterlab, jupyterlab-launcher, widgetsnbextension
packaging==20.4           # via bleach, pytest
pandocfilters==1.4.2      # via nbconvert
parso==0.7.0              # via jedi
pathspec==0.8.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.2.1          # via kedro
pluggy==0.13.1            # via kedro, pytest
poyo==0.5.0               # via cookiecutter
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.5     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.4.7          # via packaging
pyrsistent==0.16.0        # via jsonschema
pytest-cov==2.10.0        # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
pytest-mock==1.13.0       # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
pytest==5.4.3             # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in, pytest-cov, pytest-mock
python-dateutil==2.8.1    # via jupyter-client
python-json-logger==0.1.11  # via kedro
pyyaml==5.3.1             # via kedro
pyzmq==19.0.1             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.5          # via jupyter
qtpy==1.9.0               # via qtconsole
regex==2020.6.8           # via black
requests==2.24.0          # via cookiecutter
send2trash==1.5.0         # via notebook
six==1.15.0               # via bleach, jsonschema, packaging, pip-tools, python-dateutil, traitlets
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.1              # via black
toposort==1.5             # via kedro
tornado==6.0.4            # via ipykernel, jupyter-client, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
typed-ast==1.4.1          # via black
urllib3==1.25.9           # via requests
wcwidth==0.2.5            # via prompt-toolkit, pytest
webencodings==0.5.1       # via bleach
wheel==0.32.2             # via -r /Users/steeve_laquitaine/Desktop/intent/intent/src/requirements.in
whichcraft==0.6.1         # via cookiecutter
widgetsnbextension==3.5.1  # via ipywidgets
zipp==3.1.0               # via importlib-metadata
autopep8
conda
scikit-learn  
matplotlib
spacy
nltk
allennlp==1.0.0
allennlp-models==1.0.0
tensorflow==1.14.0
stanza==1.1.1
pyyaml
# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
